[FEAT] 친구요청 수락 API 구현

+++ src/main/java/com/sparta/springusersetting/domain/friends/service/FriendService.java
@@ -1,7 +1,7 @@
 package com.sparta.springusersetting.domain.friends.service;
 
 import com.sparta.springusersetting.domain.common.dto.AuthUser;
-import com.sparta.springusersetting.domain.friends.dto.responseDto.FriendRequestResponseDto;
+import com.sparta.springusersetting.domain.friends.dto.responseDto.FriendResponseDto;
 import com.sparta.springusersetting.domain.friends.entity.FriendStatus;
 import com.sparta.springusersetting.domain.friends.entity.Friends;
 import com.sparta.springusersetting.domain.friends.repository.FriendRepository;
@@ -20,7 +20,7 @@ public class FriendService {
     private final UserService userService;
     private final FriendRepository friendRepository;
 
-    public FriendRequestResponseDto requestFriend(AuthUser authUser, Long id) {
+    public FriendResponseDto requestFriend(AuthUser authUser, Long id) {
 
         User user = User.fromAuthUser(authUser);
         userService.findUser(user.getId());
@@ -39,7 +39,7 @@ public class FriendService {
         Friends friends = new Friends(user, friendUser);
         friendRepository.save(friends);
 
-        return new FriendRequestResponseDto(user, friendUser);
+        return new FriendResponseDto(friends, user, friendUser);
     }
 
     public void cancelFriend(AuthUser authUser, Long id) {
@@ -54,10 +54,42 @@ public class FriendService {
         }
 
         Friends friends = friendRequest.get();
+
+        if (!friends.getUserId().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("본인이 보낸 친구 요청만 취소할 수 있습니다.");
+        }
+
         if (friends.getFriendStatus() != FriendStatus.PENDING) {
             throw new IllegalArgumentException("이미 수락했거나 거절된 친구 요청입니다.");
         }
 
         friendRepository.delete(friends);
+    }
+
+    public FriendResponseDto acceptFriend(AuthUser authUser, Long id) {
+
+        User user = User.fromAuthUser(authUser);
+        userService.findUser(user.getId());
+
+        Optional<Friends> friendRequest = friendRepository.findByUserIdAndFriendUserId(id, user.getId());
+
+        if (friendRequest.isEmpty()) {
+            throw new IllegalArgumentException("해당 친구 요청을 찾을 수 없습니다.");
+        }
+
+        Friends friends = friendRequest.get();
+
+        if (!friends.getFriendUserId().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("본인이 받은 친구 요청만 수락할 수 있습니다.");
+        }
+
+        if (friends.getFriendStatus() == FriendStatus.ACCEPTED) {
+            throw new IllegalArgumentException("이미 수락한 친구입니다.");
+        }
+
+        friends.acceptFriend();
+        friendRepository.save(friends);
+
+        return new FriendResponseDto(friends, friends.getFriendUserId(), friends.getUserId());
     }
 }
